# Smart_School_Bus_Tracking_System/ssb-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Smart School Bus Tracking System API
  version: 1.1.0
  description: |
    API documentation for Smart School Bus Tracking System - Backend v1.1
    Refactored with normalized stops and route_stops junction table.
  contact:
    name: SSB Development Team
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
  - url: https://api.schoolbus.vn/api/v1
    description: Production server
...
paths:
  /routes:
    get:
      tags:
        - Routes
      summary: Get list of routes
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: trangThai
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteListResponse"
              example:
                success: true
                data:
                  - maTuyen: 1
                    tenTuyen: "Tuyến Quận 7 - Nhà Bè"
                    diemBatDau: "Trường Tiểu học ABC"
                    diemKetThuc: "Khu dân cư Phú Xuân"
                    thoiGianUocTinh: 45
                    origin_lat: 10.7345
                    origin_lng: 106.7212
                    dest_lat: 10.6972
                    dest_lng: 106.7041
                    polyline: null
                    trangThai: true
                pagination:
                  page: 1
                  limit: 10
                  total: 5
                  totalPages: 1
                message: "Lấy danh sách tuyến đường thành công"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Routes
      summary: Create a new route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoutePayload"
            example:
              tenTuyen: "Tuyến Quận 7 - Quận 8"
              diemBatDau: "Trường Tiểu học ABC"
              diemKetThuc: "Khu dân cư Quận 8"
              thoiGianUocTinh: 50
      responses:
        "201":
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /stops:
    get:
      tags:
        - Stops
      summary: Get list of stops
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: minLat
          in: query
          schema:
            type: number
        - name: maxLat
          in: query
          schema:
            type: number
        - name: minLng
          in: query
          schema:
            type: number
        - name: maxLng
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Stop"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  message:
                    type: string
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Stops
      summary: Create a new stop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStopPayload"
      responses:
        "201":
          description: Stop created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Stop"
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /trips:
    get:
      tags:
        - Trips
      summary: Get list of trips
      description: |
        Get paginated list of trips with filtering and sorting.
        Supports search, status filter, date filter.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: q
          in: query
          description: Search query (route name, bus plate, driver name)
          schema:
            type: string
        - name: ngayChay
          in: query
          description: Filter by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: trangThai
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [chua_khoi_hanh, dang_chay, hoan_thanh, huy]
        - name: sortBy
          in: query
          schema:
            type: string
            default: ngayChay
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Trip"
                  meta:
                    type: object
                    properties:
                      page:
                        type: integer
                      pageSize:
                        type: integer
                      total:
                        type: integer
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Trips
      summary: Create new trip
      description: Create a new trip from schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Trip created successfully
        "409":
          description: Trip already exists for this schedule and date
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
    Route:
      type: object
      properties:
        maTuyen:
          type: integer
        tenTuyen:
          type: string
        diemBatDau:
          type: string
        diemKetThuc:
          type: string
        thoiGianUocTinh:
          type: integer
        origin_lat:
          type: number
          format: float
        origin_lng:
          type: number
          format: float
        dest_lat:
          type: number
          format: float
        dest_lng:
          type: number
          format: float
        polyline:
          type: string
          nullable: true
        trangThai:
          type: boolean
        ngayTao:
          type: string
          format: date-time
        ngayCapNhat:
          type: string
          format: date-time
    Stop:
      type: object
      properties:
        maDiem:
          type: integer
        tenDiem:
          type: string
        viDo:
          type: number
          format: float
        kinhDo:
          type: number
          format: float
        address:
          type: string
          nullable: true
        scheduled_time:
          type: string
          format: time
          nullable: true
        ngayTao:
          type: string
          format: date-time
        ngayCapNhat:
          type: string
          format: date-time
...
    RouteStopDetail:
      allOf:
        - $ref: "#/components/schemas/Stop"
        - type: object
          properties:
            sequence:
              type: integer
            dwell_seconds:
              type: integer
    RouteDetail:
      allOf:
        - $ref: "#/components/schemas/Route"
        - type: object
          properties:
            stops:
              type: array
              items:
                $ref: "#/components/schemas/RouteStopDetail"
            schedules:
              type: array
              items:
                type: object
    CreateRoutePayload:
      type: object
      required:
        - tenTuyen
        - diemBatDau
        - diemKetThuc
      properties:
        tenTuyen:
          type: string
        diemBatDau:
          type: string
        diemKetThuc:
          type: string
        thoiGianUocTinh:
          type: integer
        origin_lat:
          type: number
          format: float
        origin_lng:
          type: number
          format: float
        dest_lat:
          type: number
          format: float
        dest_lng:
          type: number
          format: float
        polyline:
          type: string
        trangThai:
          type: boolean
          default: true
...
    CreateStopPayload:
      type: object
      required:
        - tenDiem
        - viDo
        - kinhDo
      properties:
        tenDiem:
          type: string
        viDo:
          type: number
          format: float
        kinhDo:
          type: number
          format: float
        address:
          type: string
        scheduled_time:
          type: string
          format: time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    RouteListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Route"
        pagination:
          $ref: "#/components/schemas/Pagination"
        message:
          type: string
    RouteDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/RouteDetail"
        message:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: "MISSING_REQUIRED_FIELDS"
              message: "Missing required fields"
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: "DUPLICATE_ROUTE_NAME"
              message: "Tên tuyến đường đã tồn tại"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "Lỗi server"

    # ═══════════════════════════════════════════════════════════════════════════
    # 🚌 TRIP SCHEMAS (M5)
    # ═══════════════════════════════════════════════════════════════════════════
...